Improved reporting for scenario outlines and data-driven tests.

+++ serenity-reports/src/test/java/net/thucydides/core/reports/integration/WhenGeneratingALargerAggregateHtmlReportSet.java
@@ -1,13 +1,14 @@
 package net.thucydides.core.reports.integration;
 
 import net.thucydides.core.issues.IssueTracking;
+import net.thucydides.core.model.TestResult;
+import net.thucydides.core.reports.OutcomeFormat;
+import net.thucydides.core.reports.TestOutcomeLoader;
+import net.thucydides.core.reports.TestOutcomes;
 import net.thucydides.core.reports.html.HtmlAggregateStoryReporter;
 import net.thucydides.core.util.EnvironmentVariables;
 import net.thucydides.core.environment.MockEnvironmentVariables;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
+import org.junit.jupiter.api.*;
 import org.mockito.MockitoAnnotations;
 
 import java.io.File;
@@ -15,27 +16,28 @@ import java.io.IOException;
 
 import static net.thucydides.core.matchers.FileMatchers.exists;
 import static net.thucydides.core.util.TestResources.directoryInClasspathCalled;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.is;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 
-public class WhenGeneratingALargerAggregateHtmlReportSet {
+class WhenGeneratingALargerAggregateHtmlReportSet {
 
     private static File outputDirectory;
     HtmlAggregateStoryReporter reporter;
 
     private static EnvironmentVariables environmentVariables = new MockEnvironmentVariables();
 
-    @BeforeClass
+    @BeforeAll
     public static void generateReports() throws IOException {
         environmentVariables.setProperty("output.formats", "json");
         environmentVariables.setProperty("report.customfields.env", "testenv");
 
     }
 
-    @AfterClass
+    @AfterAll
     public static void deleteReportDirectory() {
-        outputDirectory.delete();
+        if (outputDirectory != null) {
+            outputDirectory.delete();
+        }
     }
 
     private static File newTemporaryDirectory() throws IOException {
@@ -45,7 +47,7 @@ public class WhenGeneratingALargerAggregateHtmlReportSet {
         return createdFolder;
     }
 
-    @Before
+    @BeforeEach
     public void setupTestReporter() {
         MockitoAnnotations.initMocks(this);
     }
@@ -58,6 +60,48 @@ public class WhenGeneratingALargerAggregateHtmlReportSet {
         reporter.setOutputDirectory(outputDirectory);
         File sourceDirectory = directoryInClasspathCalled("/sample-large-site");
         reporter.generateReportsForTestResultsFrom(sourceDirectory);
-        assertThat(new File(outputDirectory, "index.html"), exists());
+        assertThat(new File(outputDirectory, "index.html")).exists();
     }
+
+    @Test
+    public void should_generate_an_aggregate_report_for_data_driven_tests() throws Exception {
+        IssueTracking issueTracking = mock(IssueTracking.class);
+        HtmlAggregateStoryReporter reporter = new HtmlAggregateStoryReporter("project", "", issueTracking, environmentVariables);
+        outputDirectory = newTemporaryDirectory();
+        reporter.setOutputDirectory(outputDirectory);
+        File sourceDirectory = directoryInClasspathCalled("/sample-data-driven-tests");
+        reporter.generateReportsForTestResultsFrom(sourceDirectory);
+        assertThat(new File(outputDirectory, "index.html")).exists();
+    }
+
+    @Nested
+    class WhenCalculatingTestResultsForAggregateReports {
+
+        TestOutcomes allTestOutcomes = TestOutcomeLoader.loadTestOutcomes().inFormat(OutcomeFormat.JSON).from(directoryInClasspathCalled("/sample-data-driven-tests"));
+
+        WhenCalculatingTestResultsForAggregateReports() throws IOException {}
+
+        @Test
+        public void should_calculate_the_total_number_of_test_outcomes_by_result() throws Exception {
+            TestOutcomes passingTestCases = allTestOutcomes.withResult(TestResult.SUCCESS);
+            TestOutcomes failingTestCases = allTestOutcomes.withResult(TestResult.FAILURE);
+
+            assertThat(passingTestCases.getScenarioCount()).isEqualTo(20);
+            assertThat(passingTestCases.getTestCaseCount()).isEqualTo(40);
+            assertThat(failingTestCases.getScenarioCount()).isEqualTo(1);
+            assertThat(failingTestCases.getTestCaseCount()).isEqualTo(3);
+        }
+
+        @Test
+        public void should_calculate_the_total_number_of_test_outcomes_by_duration() throws Exception {
+            TestOutcomes passingTestCases = allTestOutcomes.withResult(TestResult.SUCCESS);
+            TestOutcomes failingTestCases = allTestOutcomes.withResult(TestResult.FAILURE);
+
+            assertThat(passingTestCases.getScenarioCount()).isEqualTo(20);
+            assertThat(passingTestCases.getTestCaseCount()).isEqualTo(40);
+            assertThat(failingTestCases.getScenarioCount()).isEqualTo(1);
+            assertThat(failingTestCases.getTestCaseCount()).isEqualTo(3);
+        }
+    }
+
 }
